🎨 chore: 优化 chatSelectors
