+++ src/store/chat/selectors/index.ts
@@ -1,19 +1,2 @@
-import {
-  chatsMessageString,
-  currentChats,
-  currentChatsWithGuideMessage,
-  currentChatsWithHistoryConfig,
-  getFunctionMessageProps,
-  getMessageById,
-} from './chat';
-
+export { chatSelectors } from './chat';
 export { topicSelectors } from './topic';
-
-export const chatSelectors = {
-  chatsMessageString,
-  currentChats,
-  currentChatsWithGuideMessage,
-  currentChatsWithHistoryConfig,
-  getFunctionMessageProps,
-  getMessageById,
-};
